+++ src/test/static/class/reexported-classes.js
@@ -1,0 +1,1 @@
+export { ExportedClass as ReexportedClass } from './class-names.js';